Store history values in 16 bit integer so save some cache memory.

+++ RubiChess/RubiChess.h
@@ -868,7 +868,7 @@ public:
     int quietmovenum;
     int legalmovenum;
     bool onlyGoodCaptures;
-    int32_t *cmptr[CMPLIES];
+    int16_t *cmptr[CMPLIES];
 
 public:
     void SetPreferredMoves(chessposition *p);  // for quiescence move selector
@@ -902,7 +902,7 @@ enum MoveType { QUIET = 1, CAPTURE = 2, PROMOTE = 4, TACTICAL = 6, ALL = 7, EVAS
 enum RootsearchType { SinglePVSearch, MultiPVSearch };
 
 template <MoveType Mt> int CreateMovelist(chessposition *pos, chessmove* m);
-template <MoveType Mt> void evaluateMoves(chessmovelist *ml, chessposition *pos, int32_t **cmptr);
+template <MoveType Mt> void evaluateMoves(chessmovelist *ml, chessposition *pos, int16_t **cmptr);
 
 enum AttackType { FREE, OCCUPIED };
 
@@ -942,8 +942,8 @@ public:
     chessmove bestmove[MAXMULTIPV];
     int bestmovescore[MAXMULTIPV];
     uint32_t killer[2][MAXDEPTH];
-    int32_t history[2][64][64];
-    int32_t counterhistory[14][64][14*64];
+    int16_t history[2][64][64];
+    int16_t counterhistory[14][64][14*64];
     uint32_t bestFailingLow;
     Pawnhash *pwnhsh;
     repetition rp;
@@ -1013,8 +1013,8 @@ public:
     template <RootsearchType RT> int rootsearch(int alpha, int beta, int depth);
     int alphabeta(int alpha, int beta, int depth);
     int getQuiescence(int alpha, int beta, int depth);
-    void updateHistory(uint32_t code, int32_t **cmptr, int value);
-    void getCmptr(int32_t **cmptr);
+    void updateHistory(uint32_t code, int16_t **cmptr, int value);
+    void getCmptr(int16_t **cmptr);
 
 #ifdef SDEBUG
     void updatePvTable(uint32_t movecode);
